+++ x/erc20/keeper/evm_hooks_test.go
@@ -33,7 +33,7 @@ func (suite *KeeperTestSuite) TestEvmHooksRegisterERC20() {
 				suite.Require().NoError(err)
 
 				// Mint 10 tokens to suite.address (owner)
-				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10))
+				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10), false)
 				suite.Commit()
 
 				// Burn the 10 tokens of suite.address (owner)
@@ -49,7 +49,7 @@ func (suite *KeeperTestSuite) TestEvmHooksRegisterERC20() {
 			"unregistered pair",
 			func(contractAddr common.Address) {
 				// Mint 10 tokens to suite.address (owner)
-				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10))
+				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10), false)
 				suite.Commit()
 
 				// Burn the 10 tokens of suite.address (owner)
@@ -68,7 +68,7 @@ func (suite *KeeperTestSuite) TestEvmHooksRegisterERC20() {
 				suite.Require().NoError(err)
 
 				// Mint 10 tokens to suite.address (owner)
-				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10))
+				_ = suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(10), false)
 
 				stateDB := suite.StateDB()
 				logs := stateDB.Logs()