+++ x/erc20/keeper/msg_server.go
@@ -8,6 +8,9 @@ import (
 	sdkerrors "github.com/cosmos/cosmos-sdk/types/errors"
 	"github.com/ethereum/go-ethereum/accounts/abi"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/crypto"
+
+	evmtypes "github.com/tharsis/ethermint/x/evm/types"
 
 	"github.com/tharsis/evmos/x/erc20/types"
 	"github.com/tharsis/evmos/x/erc20/types/contracts"
@@ -212,6 +215,11 @@ func (k Keeper) convertCoinNativeERC20(
 		)
 	}
 
+	// Check for unexpected `appove` event in logs
+	if err := k.monitorApprovalEvent(res); err != nil {
+		return nil, err
+	}
+
 	ctx.EventManager().EmitEvents(
 		sdk.Events{
 			sdk.NewEvent(
@@ -302,6 +310,7 @@ func (k Keeper) convertERC20NativeCoin(
 //  - Send minted coins to the receiver
 //  - Check if coin balance increased by amount
 //  - Check if token balance decreased by amount
+//  - Check for unexpected `appove` event in logs
 func (k Keeper) convertERC20NativeToken(
 	ctx sdk.Context,
 	pair types.TokenPair,
@@ -369,6 +378,11 @@ func (k Keeper) convertERC20NativeToken(
 		)
 	}
 
+	// Check for unexpected `appove` event in logs
+	if err := k.monitorApprovalEvent(res); err != nil {
+		return nil, err
+	}
+
 	ctx.EventManager().EmitEvents(
 		sdk.Events{
 			sdk.NewEvent(
@@ -385,6 +399,7 @@ func (k Keeper) convertERC20NativeToken(
 	return &types.MsgConvertERC20Response{}, nil
 }
 
+// balanceOf queries an account's balance for a given ERC20 contract
 func (k Keeper) balanceOf(
 	ctx sdk.Context,
 	abi abi.ABI,
@@ -406,4 +421,24 @@ func (k Keeper) balanceOf(
 	}
 
 	return balance
+}
+
+// monitorApprovalEvent returns an error if the given transactions logs include
+// an unexpected `approve` event
+func (k Keeper) monitorApprovalEvent(res *evmtypes.MsgEthereumTxResponse) error {
+	if res == nil || len(res.Logs) == 0 {
+		return nil
+	}
+
+	logApprovalSig := []byte("Approval(address,address,uint256)")
+	logApprovalSigHash := crypto.Keccak256Hash(logApprovalSig)
+
+	for _, log := range res.Logs {
+		if log.Topics[0] == logApprovalSigHash.Hex() {
+			return sdkerrors.Wrapf(
+				types.ErrUnexpectedEvent, "unexpected approval event",
+			)
+		}
+	}
+	return nil
 }