+++ x/erc20/keeper/msg_server_test.go
@@ -175,17 +175,20 @@ func (suite *KeeperTestSuite) TestConvertECR20NativeCoin() {
 }
 
 func (suite *KeeperTestSuite) TestConvertECR20NativeERC20() {
+	var contractAddr common.Address
+
 	testCases := []struct {
-		name           string
-		mint           int64
-		burn           int64
-		malleate       func(common.Address)
-		expPass        bool
-		selfdestructed bool
+		name               string
+		mint               int64
+		burn               int64
+		malleate           func(common.Address)
+		isMaliciousDelayed bool
+		selfdestructed     bool
+		expPass            bool
 	}{
-		{"ok - sufficient funds", 100, 10, func(common.Address) {}, true, false},
-		{"ok - equal funds", 10, 10, func(common.Address) {}, true, false},
-		{"ok - equal funds", 10, 10, func(common.Address) {}, true, false},
+		{"ok - sufficient funds", 100, 10, func(common.Address) {}, false, false, true},
+		{"ok - equal funds", 10, 10, func(common.Address) {}, false, false, true},
+		{"ok - equal funds", 10, 10, func(common.Address) {}, false, false, true},
 		{
 			"ok - suicided contract",
 			10,
@@ -196,10 +199,11 @@ func (suite *KeeperTestSuite) TestConvertECR20NativeERC20() {
 				suite.Require().True(ok)
 				suite.Require().NoError(stateDb.Commit())
 			},
+			false,
 			true,
 			true,
 		},
-		{"fail - insufficient funds - callEVM", 0, 10, func(common.Address) {}, false, false},
+		{"fail - insufficient funds - callEVM", 0, 10, func(common.Address) {}, false, false, false},
 		{
 			"fail - minting disabled",
 			100,
@@ -211,13 +215,25 @@ func (suite *KeeperTestSuite) TestConvertECR20NativeERC20() {
 			},
 			false,
 			false,
+			false,
+		},
+		{
+			"fail - delayed malicious contract",
+			10,
+			10,
+			func(common.Address) {
+				contractAddr = suite.setupRegisterERC20PairMaliciousDelayed()
+			},
+			true,
+			false,
+			false,
 		},
 	}
 	for _, tc := range testCases {
 		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
 			suite.mintFeeCollector = true
 			suite.SetupTest()
-			contractAddr := suite.setupRegisterERC20Pair()
+			contractAddr = suite.setupRegisterERC20Pair()
 			suite.Require().NotNil(contractAddr)
 
 			tc.malleate(contractAddr)
@@ -231,8 +247,9 @@ func (suite *KeeperTestSuite) TestConvertECR20NativeERC20() {
 				contractAddr,
 				suite.address,
 			)
+			// FIXME: gas
 			suite.MintFeeCollector(sdk.NewCoins(sdk.NewCoin("aphoton", sdk.NewInt(int64(params.TxGas)))))
-			suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(tc.mint))
+			suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(tc.mint), tc.isMaliciousDelayed)
 			suite.Commit()
 			ctx := sdk.WrapSDKContext(suite.ctx)
 
@@ -269,29 +286,48 @@ func (suite *KeeperTestSuite) TestConvertECR20NativeERC20() {
 }
 
 func (suite *KeeperTestSuite) TestConvertCoinNativeERC20() {
+	var contractAddr common.Address
+
 	testCases := []struct {
-		name      string
-		mint      int64
-		burn      int64
-		reconvert int64
-		expPass   bool
+		name               string
+		mint               int64
+		burn               int64
+		reconvert          int64
+		malleate           func(common.Address)
+		isMaliciousDelayed bool
+		expPass            bool
 	}{
-		{"ok - sufficient funds", 100, 10, 5, true},
-		{"ok - equal funds", 10, 10, 10, true},
-		{"fail - insufficient funds", 10, 1, 5, false},
+		{"ok - sufficient funds", 100, 10, 5, func(common.Address) {}, false, true},
+		{"ok - equal funds", 10, 10, 10, func(common.Address) {}, false, true},
+		{"fail - insufficient funds", 10, 1, 5, func(common.Address) {}, false, false},
+		{
+			"fail - delayed malicious contract",
+			100,
+			10,
+			5,
+			func(common.Address) {
+				contractAddr = suite.setupRegisterERC20PairMaliciousDelayed()
+			},
+			true,
+			false,
+		},
 	}
 	for _, tc := range testCases {
 		suite.Run(fmt.Sprintf("Case %s", tc.name), func() {
 			suite.mintFeeCollector = true
 			suite.SetupTest()
-			contractAddr := suite.setupRegisterERC20Pair()
+			contractAddr = suite.setupRegisterERC20Pair()
 			suite.Require().NotNil(contractAddr)
+
+			tc.malleate(contractAddr)
+			suite.Commit()
 
 			// Precondition: Convert ERC20 to Coins
 			coinName := types.CreateDenom(contractAddr.String())
 			sender := sdk.AccAddress(suite.address.Bytes())
-			suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(tc.mint))
+			suite.MintERC20Token(contractAddr, suite.address, suite.address, big.NewInt(tc.mint), tc.isMaliciousDelayed)
 			suite.Commit()
+
 			msgConvertERC20 := types.NewMsgConvertERC20(
 				sdk.NewInt(tc.burn),
 				sender,
@@ -301,12 +337,17 @@ func (suite *KeeperTestSuite) TestConvertCoinNativeERC20() {
 
 			ctx := sdk.WrapSDKContext(suite.ctx)
 			_, err := suite.app.Erc20Keeper.ConvertERC20(ctx, msgConvertERC20)
-			suite.Require().NoError(err)
 			suite.Commit()
 			balance := suite.BalanceOf(contractAddr, suite.address)
 			cosmosBalance := suite.app.BankKeeper.GetBalance(suite.ctx, sender, coinName)
-			suite.Require().Equal(cosmosBalance.Amount, sdk.NewInt(tc.burn))
-			suite.Require().Equal(balance.(*big.Int).Int64(), big.NewInt(tc.mint-tc.burn).Int64())
+
+			if tc.isMaliciousDelayed {
+				suite.Require().Error(err)
+			} else {
+				suite.Require().NoError(err)
+				suite.Require().Equal(cosmosBalance.Amount, sdk.NewInt(tc.burn))
+				suite.Require().Equal(balance.(*big.Int).Int64(), big.NewInt(tc.mint-tc.burn).Int64())
+			}
 
 			// Convert Coins back to ERC20s
 			ctx = sdk.WrapSDKContext(suite.ctx)